# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "运维管理平台 API 文档",
        "title": "OpsHub API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "url": "http://www.swagger.io/support",
            "email": "support@swagger.io"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
...
        "/monitor/domains": {
            "get": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "DomainMonitor"
                ],
                "summary": "获取域名监控列表",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/model.DomainMonitor"
                            }
                        }
                    }
                }
            },
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "DomainMonitor"
                ],
                "summary": "创建域名监控",
                "parameters": [
                    {
                        "description": "域名监控信息",
                        "name": "monitor",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.DomainMonitor"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.DomainMonitor"
                        }
                    }
                }
            }
        },
...
        "/monitor/domains/{id}": {
            "get": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "DomainMonitor"
                ],
                "summary": "获取域名监控详情",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "域名ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.DomainMonitor"
                        }
                    }
                }
            },
            "put": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "DomainMonitor"
                ],
                "summary": "更新域名监控",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "域名ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "域名监控信息",
                        "name": "monitor",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.DomainMonitor"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.DomainMonitor"
                        }
                    }
                }
            },
            "delete": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "DomainMonitor"
                ],
                "summary": "删除域名监控",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "域名ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "success",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/monitor/domains/{id}/check": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "DomainMonitor"
                ],
                "summary": "立即检查域名",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "域名ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.DomainMonitor"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "model.DomainMonitor": {
            "type": "object",
            "properties": {
                "alertConfigId": {
                    "description": "告警配置",
                    "type": "integer"
                },
                "checkInterval": {
                    "description": "检查间隔(秒)",
                    "type": "integer"
                },
                "createdAt": {
                    "type": "string"
                },
                "domain": {
                    "description": "域名",
                    "type": "string"
                },
                "enableAlert": {
                    "description": "是否启用告警",
                    "type": "boolean"
                },
                "enableSSL": {
                    "description": "是否启用SSL检查",
                    "type": "boolean"
                },
                "id": {
                    "type": "integer"
                },
                "lastCheck": {
                    "description": "最后检查时间",
                    "type": "string"
                },
                "nextCheck": {
                    "description": "下次检查时间",
                    "type": "string"
                },
                "responseThreshold": {
                    "description": "响应时间阈值(ms)",
                    "type": "integer"
                },
                "responseTime": {
                    "description": "响应时间(ms)",
                    "type": "integer"
                },
                "sslExpiry": {
                    "description": "SSL过期时间",
                    "type": "string"
                },
                "sslExpiryDays": {
                    "description": "SSL过期提前告警天数",
                    "type": "integer"
                },
                "sslValid": {
                    "description": "SSL是否有效",
                    "type": "boolean"
                },
                "status": {
                    "description": "状态: normal, abnormal, paused, unknown",
                    "type": "string"
                },
                "updatedAt": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "Bearer": {
            "description": "Type \"Bearer\" followed by a space and JWT token.",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}